% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_choirbm_list.R
\name{agg_choirbm_list}
\alias{agg_choirbm_list}
\title{Converts a list of CBMs to a single data frame}
\usage{
agg_choirbm_list(cbm_list)
}
\arguments{
\item{cbm_list}{a list of CBMs to collapse}
}
\value{
map_df a single CBM data frame with the value column summed.
}
\description{
Takes a list of data frames where each data frame is the CBM of a patient
and the values column is a binary endorsement of the CBM segment. Then it
collapses the information to a single data frame for plotting by adding
the 'value' columns of each data frame.
}
\examples{
\dontrun{
data(validation)
cbm_list <- lapply(validation[["bodymap_regions_csv"]], string_to_map)
agg_df <- agg_choirbm_list(cbm_list)
}
}
