% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHOIRBM-package.R
\docType{package}
\name{CHOIRBM-package}
\alias{CHOIRBM}
\alias{CHOIRBM-package}
\title{CHOIRBM: Plots the CHOIR Body Map}
\description{
Collection of utility functions for visualizing
    body map data collected with the Collaborative Health Outcomes
    Information Registry.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/emcramer/CHOIRBM}
  \item Report bugs at \url{https://github.com/emcramer/CHOIRBM/issues}
}

}
\author{
\strong{Maintainer}: Eric Cramer \email{emcramer@stanford.edu} (\href{https://orcid.org/0000-0003-2085-3679}{ORCID})

Other contributors:
\itemize{
  \item Stanford University School of Medicine [copyright holder, funder]
}

}
\keyword{internal}
