% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_choirbm_chi.R
\name{comp_choirbm_chi}
\alias{comp_choirbm_chi}
\title{Compare CBM segment endorsement across categorical variable using Chi-Square}
\usage{
comp_choirbm_chi(cbm_list, ...)
}
\arguments{
\item{cbm_list}{a named list of CBMs}

\item{...}{additional parameters passed to p.adjust()}
}
\value{
a data frame with the p-values, chi statistic, and degrees of freedom
}
\description{
Compare CBM segment endorsement across categorical variable using Chi-Square
}
\examples{
\dontrun{
data(validation)
# split male and female data
male_data <- validation[validation[['gender']] == "Male", ]
male_bodymap_list <- lapply(male_data[["bodymap_regions_csv"]], string_to_map)
male_bodymap_df <- agg_choirbm_list(male_bodymap_list)
female_data <- validation[validation[['gender']] == "Female", ]
female_bodymap_list <- lapply(female_data[["bodymap_regions_csv"]], string_to_map)
female_bodymap_df <- agg_choirbm_list(female_bodymap_list)
# compare with chi square test
chi_res <- comp_choirbm_chi(
 list("male" = male_bodymap_df
 , "female" = female_bodymap_df)
 , method = 'bonferroni'
)
}
}
