% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_choirbm_glm.R
\name{comp_choirbm_glm}
\alias{comp_choirbm_glm}
\title{Examine the effect of a continuous variable on CBM location endorsement}
\usage{
comp_choirbm_glm(in_df, comp_var, ...)
}
\arguments{
\item{in_df}{a data.frame with at least one column for the CBM as a
delimited string, and another column as the continuous variable for
modeling.}

\item{comp_var}{the name of the variable to model as a string.}

\item{...}{additional parameters passed to glm.}
}
\value{
a data.frame with the following columns: id, term, estimate,
std.error, statistic, p.value. Each row is the result of one glm
using the continuous variable to predict CBM location endorsement.
}
\description{
Examine the effect of a continuous variable on CBM location endorsement
}
\examples{
\dontrun{
data(validation)
set.seed(123)
sampled_data <- validation[sample(1:nrow(validation), 100, replace = FALSE),]
model_ouput <- comp_choirbm_glm(sampled_data, "age")
}

}
