% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_choirbm_ztest.R
\name{comp_choirbm_ztest}
\alias{comp_choirbm_ztest}
\title{Compare CBM segment endorsement across categorical variable using z-test(s)}
\usage{
comp_choirbm_ztest(cbm_list, tail = "two", p.method = "bonferroni")
}
\arguments{
\item{cbm_list}{a named list of CBMs}

\item{tail}{whether to do a single or two tailed z test}

\item{...}{additional parameters passed to p.adjust()}
}
\value{
a data frame with the p-values and z statistic
}
\description{
Compare CBM segment endorsement across categorical variable using z-test(s)
}
\examples{
library(CHOIRBM)
# isolate and process male data
male_data <- validation[validation[["gender"]] == "Male", ]
# isolate and process female data
female_data <- validation[validation[["gender"]] == "Female", ]
comp_choirbm_ztest(list( "male" = male_data, "female" = female_data), tail = "two")
}
