% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_concurrence.R
\name{comp_concurrence}
\alias{comp_concurrence}
\title{Calculate the concurrence between locations on the CBM}
\usage{
comp_concurrence(df)
}
\arguments{
\item{df}{a data.frame with the CBMs as delimited strings in a single column.}
}
\value{
a data.frame with every combination of CBM locations and the
number of times those locations occur together (the "concurrence").
}
\description{
Calculates the raw number of times two locations on the CBM are endorsed
together in a data set.
}
\examples{
\dontrun{
set.seed(123)
sampled_data <- validation[sample(1:nrow(validation), 100, replace = FALSE),]
colnames(sampled_data)[5] <- "bodymap"
con_mat <- comp_concurrence(sampled_data)
}
}
