% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_bodymaps.R
\name{convert_bodymaps}
\alias{convert_bodymaps}
\title{convert_bodymaps
Function to convert multiple bodymaps}
\usage{
convert_bodymaps(f_maps)
}
\arguments{
\item{f_maps}{a character vector where each string is a CHOIR bodymap
in csv form}
}
\value{
a character vector of bodymaps using the male CHOIR bodymap
numberings as a standard. Each bodymap is in csv form
}
\description{
convert_bodymaps
Function to convert multiple bodymaps
}
\examples{
exampledata <- data.frame(
    GENDER = as.character(c("Male", "Female", "Female")),
    BODYMAP_CSV = as.character(c("112,125","112,113","128,117"))
 )
convert_bodymaps(
    as.character(
        exampledata$BODYMAP_CSV[exampledata$GENDER == 'Female']
     )
 )
}
