% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_example_data.R
\name{gen_example_data}
\alias{gen_example_data}
\title{Generate Simple Example Data}
\usage{
gen_example_data(seed = 123)
}
\arguments{
\item{seed}{integer to seed the random number generator}
}
\value{
values data.frame
}
\description{
Creates a data frame with CHOIR Body Map segment IDs and a randomly associated value. Also adds grouping information for facetting while plotting.
}
\examples{

}
