% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_areas.R
\name{num_areas}
\alias{num_areas}
\title{Count the number of areas indicated in a CBM}
\usage{
num_areas(cbm_str, delim = ",")
}
\arguments{
\item{cbm_str}{a delimited string of 3-digit codes indicating CBM areas.}

\item{delim}{the delimiter character, defaults to a comma.}
}
\value{
nareas
}
\description{
Counts the number of areas a person endorses/indicates on their CHOIR
Body Map.
}
\examples{
cbm_str <- c("101,102,103,104")
num_areas(cbm_str, ",")
}
