% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_female_choirbm.R
\name{plot_female_choirbm}
\alias{plot_female_choirbm}
\title{Plot the male CHOIR Body Map}
\usage{
plot_female_choirbm(df, value)
}
\arguments{
\item{df}{data.frame}

\item{value}{string}
}
\value{
ggrob
}
\description{
Creates a new plot of the front and back of the female CHOIR body map.
}
\examples{
cbm_df <- gen_example_data()
plot_female_choirbm(cbm_df, "value")

}
