% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_male_choirbm.R
\name{plot_male_choirbm}
\alias{plot_male_choirbm}
\title{Plot the male CHOIR Body Map}
\usage{
plot_male_choirbm(df, value)
}
\arguments{
\item{df}{data.frame}

\item{value}{string}
}
\value{
ggrob
}
\description{
Creates a new plot of the male CHOIR body map, with the option to sweitch between front and back.
}
\examples{
cbm_df <- gen_example_data()
plot_male_choirbm(cbm_df, "value")

}
