% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_male_back_choirbm.R,
%   R/plot_male_front_choirbm.R
\name{plot_male_front_choirbm}
\alias{plot_male_front_choirbm}
\title{Plot the back part of the male CHOIR Body Map}
\usage{
plot_male_front_choirbm(df)

plot_male_front_choirbm(df)
}
\arguments{
\item{df}{data.frame}
}
\value{


ggrob
}
\description{
Creates a new plot of the back side of the male CHOIR body map.

Creates a new plot of the front side of the male CHOIR body map.
}
