% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHOIRBM-package.R
\docType{package}
\name{CHOIRBM-package}
\alias{CHOIRBM}
\alias{CHOIRBM-package}
\title{CHOIRBM: Plots the CHOIR Body Map}
\description{
Colleciton of utility functions for visualizing data collected with the Collaborative Health Outcomes Information Registry.
}
\author{
\strong{Maintainer}: Eric Cramer \email{emcramer@stanford.edu} (\href{https://orcid.org/0000-0003-2085-3679}{ORCID})

}
\keyword{internal}
