% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_choirbm_ztest.R
\name{comp_choirbm_ztest}
\alias{comp_choirbm_ztest}
\title{Compare CBM segment endorsement across categorical variable using z-test(s)}
\usage{
comp_choirbm_ztest(df1, df2, tail = "two", p.method = "bonferroni")
}
\arguments{
\item{cbm_list}{a named list of CBMs}

\item{...}{additional parameters passed to p.adjust()}
}
\value{
a data frame with the p-values and z statistic
}
\description{
Compare CBM segment endorsement across categorical variable using z-test(s)
}
